api/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
python-dotenv>=1.0.0
anthropic>=0.42.0
